api/requirements.txt
# Azure Functions Python dependencies
azure-functions>=1.17.0

# GRRI-MAC Framework dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Database (Azure SQL)
pyodbc>=5.0.0
